Added blocking ConsoleServiceHost
I copied over ConsoleRunHost from topshelf and modified it to work with
DotNetCore.WindowsService library. The console service host blocks until
the user hits the CancelKey(Ctrl+Break or Ctrl+C). The ExampleService.cs
is updated to reflect the library update. The update allows for
debugging of services which are push based rather than pull based.